+++ src/entry.ts
@@ -1,5 +1,5 @@
 import './app.css';
-import App from './App.svelte';
+import App from './routes/App.svelte';
 
 if (chrome.devtools.panels.themeName === 'dark') {
 	document.body.classList.add('dark');
