+++ src/lib/nodes/Collapse.svelte
