+++ src/lib/nodes/SearchTerm.svelte
