+++ src/lib/nodes/Text.svelte
