+++ src/lib/panel/CollapsableValue.svelte
