+++ src/lib/panel/ComponentView.svelte
