+++ src/lib/profiler/Frame.svelte
