+++ src/lib/toolbar/VisibilityButton.svelte
